refactor(agent): extract viewer runtime
